[issue-45]modify clearly components' structure
