+++ ui/pdk-ui/src/components/Table/SinkTable.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import { sinkListElem } from '../ElemInterface/ElementsInterface';
-import { SINK_URL } from '../defineUrl';
+import { sinkListElem } from '../../ElemInterface/ElementsInterface';
+import { SINK_URL } from '../../defineUrl';
 
 //import DeleteRequest from './DeleteRequest'
 
@@ -42,8 +42,8 @@ class SinkTable extends Component<SinkTableProps> {
 								<th scope="row">{idx}</th>
 								<td>{sink.name}</td>
 								<td>{sink.id}</td>
-                                <td>{sink.location}</td>
-                                <td>{sink.ip}</td>
+								<td>{sink.location}</td>
+								<td>{sink.ip}</td>
 								<td>
 									<button
 										className="btn btn-default btn-sm"
